+++ apps/wasm/mina-berkeley/src/payment.rs
@@ -71,81 +71,3 @@ impl Payment {
         }
     }
 }
-
-#[wasm_bindgen(js_name = accountInfo)]
-pub async fn account_info(pub_key: String) -> Result<String, JsError> {
-    let client = reqwest::Client::new();
-    let query: String = format!(
-        "query MyQuery {{  
-            account(publicKey: \"{}\") {{ balance {{ total }} inferredNonce locked  }}
-        }}",
-        pub_key
-    );
-    let res = client
-        .post("http://localhost:3085/graphql")
-        .json(&serde_json::json!({ "query": query }))
-        .send()
-        .await
-        .unwrap();
-    let resp_text = res.text().await.unwrap();
-    let resp: serde_json::Value = serde_json::from_str(&resp_text).unwrap();
-    Ok(resp.to_string())
-}
-
-#[wasm_bindgen(js_name = lockAccount)]
-pub async fn lock_account(pub_key: String) -> Result<String, JsError> {
-    let client = reqwest::Client::new();
-    let query: String = format!(
-        "mutation MyMutation {{  
-            lockAccount(input: {{publicKey: \"{}\"}}) {{    account {{      locked    }}  }}
-        }}",
-        pub_key
-    );
-    let res = client
-        .post("http://localhost:3085/graphql")
-        .json(&serde_json::json!({ "query": query }))
-        .send()
-        .await
-        .unwrap();
-    let resp_text = res.text().await.unwrap();
-    let resp: serde_json::Value = serde_json::from_str(&resp_text).unwrap();
-    Ok(resp.to_string())
-}
-
-#[wasm_bindgen(js_name = unlockAccount)]
-pub async fn unlock_account(pub_key: String, password: String) -> Result<String, JsError> {
-    let client = reqwest::Client::new();
-    let query: String = format!(
-        "mutation MyMutation {{  
-            unlockAccount(input: {{publicKey: \"{}\", password: \"{}\"}}) {{    account {{      locked    }}  }}
-        }}", pub_key, password
-    );
-    let res = client
-        .post("http://localhost:3085/graphql")
-        .json(&serde_json::json!({ "query": query }))
-        .send()
-        .await
-        .unwrap();
-    let resp_text = res.text().await.unwrap();
-    let resp: serde_json::Value = serde_json::from_str(&resp_text).unwrap();
-    Ok(resp.to_string())
-}
-
-#[wasm_bindgen(js_name = sendPayment)]
-pub async fn send_payment(payment: Payment) -> Result<String, JsError> {
-    let client = reqwest::Client::new();
-    let query: String = format!(
-        "mutation MyMutation {{  
-            sendPayment(input: {{fee: \"{}\", amount: \"{}\", to: \"{}\", from: \"{}\"}}) {{    payment {{      id    }}  }}
-        }}", payment.fee_u64()?, payment.amount_u64()?, payment.to(), payment.from()
-    );
-    let res = client
-        .post("http://localhost:3085/graphql")
-        .json(&serde_json::json!({ "query": query }))
-        .send()
-        .await
-        .unwrap();
-    let resp_text = res.text().await.unwrap();
-    let resp: serde_json::Value = serde_json::from_str(&resp_text).unwrap();
-    Ok(resp.to_string())
-}
